add more rpcs to api  and add 3 DB transaction to create genres and studios and languages also change some models

+++ api/rpc_get_all_genres.go
@@ -50,11 +50,11 @@ func (server *Server) GetAllGenres(ctx context.Context, req *pb.GetAllGenresRequ
 }
 
 func validateGetAllGenresRequest(req *pb.GetAllGenresRequest) (violations []*errdetails.BadRequest_FieldViolation) {
-	if err := utils.ValidateInt(req.GetPageNumber()); err != nil {
+	if err := utils.ValidateInt(int64(req.GetPageNumber())); err != nil {
 		violations = append(violations, fieldViolation("pageNumber", err))
 	}
 
-	if err := utils.ValidateInt(req.GetPageSize()); err != nil {
+	if err := utils.ValidateInt(int64(req.GetPageSize())); err != nil {
 		violations = append(violations, fieldViolation("pageSize", err))
 	}
